Edit and Delete actions for the posts made user specific

+++ Tweetbook/Controllers/V1/PostController.cs
@@ -9,6 +9,7 @@ using Tweetbook.Contracts.V1;
 using Tweetbook.Contracts.V1.Requests;
 using Tweetbook.Contracts.V1.Responses;
 using Tweetbook.Domain;
+using Tweetbook.Extensions;
 using Tweetbook.Services;
 
 namespace Tweetbook.Controllers.V1
@@ -31,6 +32,13 @@ namespace Tweetbook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<ActionResult> Update([FromRoute]Guid postId, [FromBody]UpdatePostRequest request)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new {error = "You do not own this post!"});
+            }
+
             var post = new Post {Id = postId, Name = request.Name};
             var isUpdated = await _postService.UpdatePostAsync(post);
 
@@ -53,7 +61,11 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post {Name = postRequest.Name};
+            var post = new Post
+            {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
             
             await _postService.CreatePostAsync(post);
 
@@ -68,6 +80,13 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post!" });
+            }
+
             var isDeleted = await _postService.DeletePostAsync(postId);
             if (!isDeleted)
                 return NotFound();
